--- a/tests/control_data/videos_data/SceneWithMultipleCallsWithNFlag.json
+++ b/tests/control_data/videos_data/SceneWithMultipleCallsWithNFlag.json
@@ -2,11 +2,11 @@
     "name": "SceneWithMultipleCallsWithNFlag",
     "movie_metadata": {
         "codec_name": "h264",
-        "width": 1920,
-        "height": 1080,
-        "avg_frame_rate": "60/1",
+        "width": 854,
+        "height": 480,
+        "avg_frame_rate": "15/1",
         "duration": "4.000000",
-        "nb_frames": "240"
+        "nb_frames": "60"
     },
     "section_dir_layout": [],
     "section_index": []