… out.txt